webkit-patch should auto-install the keyring python module https://bugs.webkit.org/show_bug.cgi?id=140547
Reviewed by David Kilzer.

Auto-install the keyring python module. This is the module that
webkit-patch uses for cross-platform access to the keyring
service.

Without this module, if the password if not in the keychain,
webkit-patch will keep asking for your username / password and
never remember them.

* Scripts/webkitpy/common/net/credentials.py:
* Scripts/webkitpy/thirdparty/__init__.py:
(AutoinstallImportHook.find_module):
(AutoinstallImportHook._install_keyring):


Identifier: 158524@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178589 268f45cc-cd09-0410-ab3c-d52691b4dbfc