Out of bounds write in canvas.toDataURL https://bugs.webkit.org/show_bug.cgi?id=140594 <rdar://problem/19449135>
Reviewed by Alexey Proskuryakov.

Source/WebCore:

In the case where we have a canvas object that does
not have premultiplied alpha (an option you can select
when using WebGL) we have to multiply out the alpha when
converting to JPEG via toDataURL.

For this we created a buffer, but were not accurately
resizing it before flattening the alpha.

Test: fast/canvas/webgl/toDataURL-unpremultipliedAlpha.html

* platform/graphics/cg/ImageBufferCG.cpp:
(WebCore::ImageDataToDataURL): Call resize once we've
determined we have enough space.

LayoutTests:

Creates a WebGL context that does not have
premultiplied alpha, fills it with 50% transparent white,
and attempts to convert the canvas to a JPEG data URL. This
exercises the code path that was not accurately
allocating data (to flatten the alpha).

* fast/canvas/webgl/toDataURL-unpremultipliedAlpha-expected.txt: Added.
* fast/canvas/webgl/toDataURL-unpremultipliedAlpha.html: Added.


Identifier: 158570@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178637 268f45cc-cd09-0410-ab3c-d52691b4dbfc