EdenCollections unnecessarily visit SmallStrings https://bugs.webkit.org/show_bug.cgi?id=140762
Reviewed by Geoffrey Garen.

* heap/Heap.cpp:
(JSC::Heap::copyBackingStores): Also added a GCPhase for copying
backing stores, which is a significant portion of garbage collection.
(JSC::Heap::visitSmallStrings): Check to see if we need to visit
SmallStrings based on the collection type.
* runtime/SmallStrings.cpp:
(JSC::SmallStrings::SmallStrings):
(JSC::SmallStrings::visitStrongReferences): Set the fact that we have
visited the SmallStrings since the last modification.
* runtime/SmallStrings.h:
(JSC::SmallStrings::needsToBeVisited): If we're doing a
FullCollection, we need to visit. Otherwise, it depends on whether
we've been visited since the last modification/allocation.


Identifier: 158799@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@178984 268f45cc-cd09-0410-ab3c-d52691b4dbfc