Perf dashboard always assigns the result of A/B testing with build request 1 https://bugs.webkit.org/show_bug.cgi?id=140382
Reviewed by Darin Adler.

The bug was caused by the expression array_get($report, 'jobId') or array_get($report, 'buildRequest')
which always evaluated to 1 when the report contained jobId. Fixed the bug by cascading array_get instead.

Also fixed a typo as well as a bug that reports were never associated with builds.

* public/include/report-processor.php:
(ReportProcessor::process): Don't use "or" to find the non-null value since that always evaluates to 1
instead of the first non-null value.
(ReportProcessor::resolve_build_id): Fixed the typo by adding the missing "$this->".
(ReportProcessor::commit): Associate the report with the corresponding build as intended.


Identifier: 158825@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179037 268f45cc-cd09-0410-ab3c-d52691b4dbfc