Get rid of invalidSelectorVector, use Bison's error recovery instead https://bugs.webkit.org/show_bug.cgi?id=141147
Reviewed by Darin Adler.

* css/CSSGrammar.y.in:
Instead of reducing a null selector, we can use a real parsing error
to get out of invalid selector endings.

When that happens, Bison will pop the stack until it can reduce any
valid error recovery rules.

The problem is to make sure there is no floating values because
none of the reduce block between the error and the recovery would
be executed.

In this case, "nth_selector_ending" is a non-recursive production of
the NTHCHILDFUNCTIONS. In turn, NTHCHILDFUNCTIONS are productions
of the non-recursive "pseudo". "pseudo" is only used as a trivial
production of "specifier". "specifier" is only used by "specifier_list".

"specifier_list" has error recovery code -> no production could have
generated a floating values between "specifier_list" and "nth_selector_ending".


Identifier: 159162@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@179485 268f45cc-cd09-0410-ab3c-d52691b4dbfc